\name{porcelainporcupines}
\alias{porcelainporcupines}
\title{porcelain_porcupines}
\usage{
porcelain_porcupine(name, n, type = c("discrete", "continuous"))
}
\description{
Generates album cover inspired color palettes.
}
\examples{
porcelain_porcupine("Landmark", 6)
porcelain_porcupine("WarmGlow", 6)
porcelain_porcupine("Bambi", 6)
porcelain_porcupine("LP3", 4)
porcelain_porcupine("Dreamland", 6)
porcelain_porcupine("HTBAHB", 6)
porcelain_porcupine("ZABA", 5)
porcelain_porcupine("Loser", 4)
porcelain_porcupine("GA1", 8)
}
